(#8099) Add delayed_job init script
This init script works for Enterprise Linux systems and manages the
delayed_job worker process.

This should work with Puppet Enterprise as well as the open source
version of the Puppet Dashboard.  To use with Puppet Enterprise, simply
set DASHBOARD_ROOT in /etc/sysconfig/delayed_job to the directory you
have Puppet Dashboard installed into.

If you're running out of a git clone, the following sysconfig script
works:

    # /etc/sysconfig/delayed_job
    # DASHBOARD_ROOT=/opt/puppet/share/puppet-dashboard
    DASHBOARD_ROOT="/usr/local/src/puppet-dashboard"
    CPUS=2
    # These are a hack for running dashboard from source
    export PATH=/opt/puppet/bin:${PATH}
    export ENVPUPPET_BASEDIR=/usr/local/src
    eval $(/usr/local/bin/envpuppet)

The name of the init script may be changed.  If the package maintainer would
like to rename the service, simply change the $name variable at the top of the
init script.